+++ adagucserverEC/GenericDataWarper/gdwDrawTriangle.h
@@ -4,6 +4,6 @@
 #define GDW_DRAWTRIANGLE_UTILS_H
 
 template <typename T>
-int gdwDrawTriangle(int *xP, int *yP, T value, int destWidth, int destHeight, void *settings, void (*drawFunction)(int, int, T, void *settings, void *genericDataWarper), void *genericDataWarper,
+int gdwDrawTriangle(double *xP, double *yP, T value, int destWidth, int destHeight, void *settings, void (*drawFunction)(int, int, T, void *settings, void *genericDataWarper), void *genericDataWarper,
                     bool aOrB);
 #endif
